requirements.txt
aiofiles==24.1.0
fastapi==0.115.5
openai==1.55.3
pydantic==2.10.2
python-dotenv==1.0.1
uvicorn==0.30.0
websockets==11.0.3
argparse==1.4.0
